+++ src/Cdn/Cdn/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Fixed null reference exception and typos in `New-AzFrontDoorCdnRule` cmdlet
 
 ## Version 1.8.0
 * Fixed mandatory parameters issue in `Get-AzCdnEndpointResourceUsage` cmdlet