+++ src/FrontDoor/FrontDoor/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* New-AzFrontDoorRulesEngineActionObject: Allow rule engine action creation without RouteConfigurationOverride 
+* New-AzFrontDoorRulesEngineActionObject: Fix DynamicCompression parameter being ignored issue
 
 ## Version 1.8.0
 * Allowed Enable-AzFrontDoorCustomDomainHttps's SecretVersion parameter to be optional to support bring-your-own-certificate auto-rotation